Scaling on files fixed in Phase 3.5
As a sidenote, I noticed that the way some objects are sized, are different from the ones in the game. For my resizings, i matched the textures to the same aspect ratio of those textures. There are textures that are currently fine in both Phase 3, and 3.5 that have different scaling than what was in the game originally. If panda scale's those automatically to the correct size, cool. If it don't, someone should compare and fix textures that don't currently match game dimensions. (By game dimensions, I don't mean the EXACT resolution, just aspect ratio.)