Palette Organization - Phase 4
These past 3 days I've spent a ton of my time hunting down all of the unpalletized textures and organizing them together with their corresponding palettes in the PAL_ME folder for Phase 4.

This should be entirely every palette available in Phase 4 for anyone to jump in and compile together into a high res and brand new palette with a fresh coat of paint!

Thanks to Toonerz for helping out with a lot of stuff as I was doing all of this the past couple of days.